Module 3: fixes for Module 1 and Module 2
