+++ src/App.js
@@ -1,24 +1,36 @@
 import logo from './logo.svg';
 import './App.css';
 
+import Counter from './components/Counter';
+import SearchBar from './components/SearchBar';
+import GenreList from './components/GenreList';
+import MoviesList from './components/MoviesList';
+
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <>
+      <div className="App">
+        <header className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <p>
+            Edit <code>src/App.js</code> and save to reload.
+          </p>
+          <a
+            className="App-link"
+            href="https://reactjs.org"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            Learn React
+          </a>
+        </header>
+      </div>
+
+      <Counter/>
+      <SearchBar/>
+      <GenreList/>
+      <MoviesList/>
+    </>
   );
 }
