+++ src/components/MovieTile.jsx
