+++ src/constants.js
@@ -1,0 +1,5 @@
+const GenreList = ['All', 'Documentary', 'Comedy', 'Horror', 'Crime'];
+
+export {
+    GenreList
+}
