+++ src/index.js
@@ -3,19 +3,11 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import Counter from './components/Counter';
-import SearchBar from './components/SearchBar';
-import GenreList from './components/GenreList';
-import MoviesList from './components/MoviesList';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    {/* <App /> */}
-    <Counter/>
-    <SearchBar/>
-    <GenreList/>
-    <MoviesList/>
+    <App />
   </React.StrictMode>
 );
